User sind vorhanden und können sich einloggen. Ebenso existieren nun Kommentare, die an Posts hängen (also ein Post kann mehrere Kommentare haben). Auch wird zur Authentifizierung ein JWT-Token genutzt

+++ src/main/java/com/micromata/webengineering/demo/post/PostService.java
@@ -2,7 +2,10 @@ package com.micromata.webengineering.demo.post;
 
 import com.micromata.webengineering.demo.user.User;
 import com.micromata.webengineering.demo.user.UserRepository;
+import com.micromata.webengineering.demo.user.UserService;
 import org.apache.tomcat.jni.Time;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,14 +20,17 @@ import java.util.List;
  */
 @Service
 public class PostService {
-    //Wir wollen nun persistiert speichern, also keine Liste mehr
-    //private List<Post> posts = new LinkedList<>();
+    private static final Logger LOG = LoggerFactory.getLogger(PostService.class);
+
 
     @Autowired
     private PostRepository repository;
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private UserService userService;
 
 
     /**
@@ -33,6 +39,10 @@ public class PostService {
      * @return post list
      */
     public Iterable<Post> getPosts() {
+        User currentUser =  userService.getCurrentUser();
+        LOG.info("Current User {}"+currentUser);
+
+
         return repository.findAll();
     }
 
@@ -43,6 +53,7 @@ public class PostService {
      * @param post the post title.
      */
     public void addPost(Post post) {
+        LOG.info("Adding post. user={}, title={}", userService.getCurrentUser().getEmail(), post.getTitle());
 
         //Wir suchen den User (im moment ja nur einer vorhanden) und adden ihn als Author zum Post. Erst danach wird
         //persistiert
@@ -53,6 +64,7 @@ public class PostService {
     }
 
     public Post getPost(Long id){
+        LOG.info("Retrieving post. user={}, id={}", userService.getCurrentUser().getEmail(), id);
         /*
         for(Post p : posts){
             if(p.getId() == id){
@@ -61,17 +73,24 @@ public class PostService {
         }
         return null;
         */
+
+
         return repository.findOne(id);
     }
 
     public void deletePost(Long id){
-        /*
-        for(Post p : posts) {
-            if (p.getId() == id) {
-                posts.remove(p);
-            }
+        //Wir schauen, ob der aktuell eingeloggte User berechtigt ist, einen Post zu loeschen
+        Post post = repository.findOne(id);
+
+        //Ist der aktuell eingeloggte User nicht der Author des Posts: Exception werfen!
+        //Den aktuellen user erhalten wir über das JWT-Token
+        if(!post.getAuthor().equals(userService.getCurrentUser())){
+            LOG.info("Deleting post not allowed. user={}, id={}", userService.getCurrentUser().getEmail(), id);
+            throw new IllegalStateException("User not allowed to delete post");
         }
-        */
+        LOG.info("Deleting post. user={}, id={}", userService.getCurrentUser().getEmail(), id);
+
+        //Im positiven Fall (akt User ist Author): Erlaube den delete
         repository.delete(id);
     }
 }
